dialite: windows no need for extra file: more portable
